Reorganize package structure and place MainActivity within app module
