+++ mastadonApp/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="com.example.presentation.MainActivity"
+            android:name="com.matrix159.mastadonclone.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
