--- mastadonApp/src/main/java/com/example/mastadonclone/MastadonApplication.kt
@@ -1,16 +1,0 @@
-package com.example.mastadonclone
-
-import android.app.Application
-import timber.log.Timber
-
-class MastadonApplication: Application() {
-
-  override fun onCreate() {
-    super.onCreate()
-    if (BuildConfig.DEBUG) {
-      Timber.plant(Timber.DebugTree())
-    } else {
-      //plant(CrashReportingTree())
-    }
-  }
-}
