+++ mastadonApp/src/main/java/com/matrix159/mastadonclone/MainActivity.kt
@@ -1,0 +1,15 @@
+package com.matrix159.mastadonclone
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import com.matrix159.mastadonclone.presentation.ui.MastadonApp
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            MastadonApp()
+        }
+    }
+}
