+++ mastadonApp/src/main/java/com/matrix159/mastadonclone/MastadonApplication.kt
@@ -1,0 +1,16 @@
+package com.matrix159.mastadonclone
+
+import android.app.Application
+import timber.log.Timber
+
+class MastadonApplication: Application() {
+
+  override fun onCreate() {
+    super.onCreate()
+    if (BuildConfig.DEBUG) {
+      Timber.plant(Timber.DebugTree())
+    } else {
+      //plant(CrashReportingTree())
+    }
+  }
+}
