--- presentation/src/main/java/com/example/presentation/ui/theme/Shape.kt
@@ -1,7 +1,0 @@
-package com.example.presentation.ui.theme
-
-//val Shapes = Shapes(
-//  small = RoundedCornerShape(4.dp),
-//  medium = RoundedCornerShape(4.dp),
-//  large = RoundedCornerShape(0.dp)
-//)
