--- presentation/src/main/java/com/example/presentation/ui/theme/Type.kt
@@ -1,19 +1,0 @@
-package com.example.presentation.ui.theme
-
-import androidx.compose.material3.Typography
-
-// Set of Material typography styles to start with
-val Typography = Typography(
-  /* Other default text styles to override
-  button = TextStyle(
-      fontFamily = FontFamily.Default,
-      fontWeight = FontWeight.W500,
-      fontSize = 14.sp
-  ),
-  caption = TextStyle(
-      fontFamily = FontFamily.Default,
-      fontWeight = FontWeight.Normal,
-      fontSize = 12.sp
-  )
-  */
-)
