+++ presentation/src/main/java/com/matrix159/mastadonclone/presentation/ui/theme/Color.kt
@@ -1,0 +1,68 @@
+package com.matrix159.mastadonclone.presentation.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val md_theme_light_primary = Color(0xFF4443EB)
+val md_theme_light_onPrimary = Color(0xFFFFFFFF)
+val md_theme_light_primaryContainer = Color(0xFFE1E0FF)
+val md_theme_light_onPrimaryContainer = Color(0xFF07006C)
+val md_theme_light_secondary = Color(0xFF3D5BA9)
+val md_theme_light_onSecondary = Color(0xFFFFFFFF)
+val md_theme_light_secondaryContainer = Color(0xFFDBE1FF)
+val md_theme_light_onSecondaryContainer = Color(0xFF00184A)
+val md_theme_light_tertiary = Color(0xFF5254AE)
+val md_theme_light_onTertiary = Color(0xFFFFFFFF)
+val md_theme_light_tertiaryContainer = Color(0xFFE1DFFF)
+val md_theme_light_onTertiaryContainer = Color(0xFF08016B)
+val md_theme_light_error = Color(0xFFBA1A1A)
+val md_theme_light_errorContainer = Color(0xFFFFDAD6)
+val md_theme_light_onError = Color(0xFFFFFFFF)
+val md_theme_light_onErrorContainer = Color(0xFF410002)
+val md_theme_light_background = Color(0xFFFFFBFF)
+val md_theme_light_onBackground = Color(0xFF1C1B1F)
+val md_theme_light_surface = Color(0xFFFFFBFF)
+val md_theme_light_onSurface = Color(0xFF1C1B1F)
+val md_theme_light_surfaceVariant = Color(0xFFE4E1EC)
+val md_theme_light_onSurfaceVariant = Color(0xFF47464F)
+val md_theme_light_outline = Color(0xFF777680)
+val md_theme_light_inverseOnSurface = Color(0xFFF3EFF4)
+val md_theme_light_inverseSurface = Color(0xFF313034)
+val md_theme_light_inversePrimary = Color(0xFFC1C1FF)
+val md_theme_light_shadow = Color(0xFF000000)
+val md_theme_light_surfaceTint = Color(0xFF4443EB)
+val md_theme_light_outlineVariant = Color(0xFFC8C5D0)
+val md_theme_light_scrim = Color(0xFF000000)
+
+val md_theme_dark_primary = Color(0xFFC1C1FF)
+val md_theme_dark_onPrimary = Color(0xFF1200A9)
+val md_theme_dark_primaryContainer = Color(0xFF2820D4)
+val md_theme_dark_onPrimaryContainer = Color(0xFFE1E0FF)
+val md_theme_dark_secondary = Color(0xFFB3C5FF)
+val md_theme_dark_onSecondary = Color(0xFF002A76)
+val md_theme_dark_secondaryContainer = Color(0xFF214290)
+val md_theme_dark_onSecondaryContainer = Color(0xFFDBE1FF)
+val md_theme_dark_tertiary = Color(0xFFC1C1FF)
+val md_theme_dark_onTertiary = Color(0xFF22227E)
+val md_theme_dark_tertiaryContainer = Color(0xFF3A3B95)
+val md_theme_dark_onTertiaryContainer = Color(0xFFE1DFFF)
+val md_theme_dark_error = Color(0xFFFFB4AB)
+val md_theme_dark_errorContainer = Color(0xFF93000A)
+val md_theme_dark_onError = Color(0xFF690005)
+val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
+val md_theme_dark_background = Color(0xFF1C1B1F)
+val md_theme_dark_onBackground = Color(0xFFE5E1E6)
+val md_theme_dark_surface = Color(0xFF1C1B1F)
+val md_theme_dark_onSurface = Color(0xFFE5E1E6)
+val md_theme_dark_surfaceVariant = Color(0xFF47464F)
+val md_theme_dark_onSurfaceVariant = Color(0xFFC8C5D0)
+val md_theme_dark_outline = Color(0xFF918F9A)
+val md_theme_dark_inverseOnSurface = Color(0xFF1C1B1F)
+val md_theme_dark_inverseSurface = Color(0xFFE5E1E6)
+val md_theme_dark_inversePrimary = Color(0xFF4443EB)
+val md_theme_dark_shadow = Color(0xFF000000)
+val md_theme_dark_surfaceTint = Color(0xFFC1C1FF)
+val md_theme_dark_outlineVariant = Color(0xFF47464F)
+val md_theme_dark_scrim = Color(0xFF000000)
+
+
+val seed = Color(0xFF595AFF)
