+++ presentation/src/main/java/com/matrix159/mastadonclone/presentation/ui/theme/Shape.kt
@@ -1,0 +1,7 @@
+package com.matrix159.mastadonclone.presentation.ui.theme
+
+//val Shapes = Shapes(
+//  small = RoundedCornerShape(4.dp),
+//  medium = RoundedCornerShape(4.dp),
+//  large = RoundedCornerShape(0.dp)
+//)
