+++ presentation/src/main/java/com/matrix159/mastadonclone/presentation/ui/theme/Type.kt
@@ -1,0 +1,19 @@
+package com.matrix159.mastadonclone.presentation.ui.theme
+
+import androidx.compose.material3.Typography
+
+// Set of Material typography styles to start with
+val Typography = Typography(
+  /* Other default text styles to override
+  button = TextStyle(
+      fontFamily = FontFamily.Default,
+      fontWeight = FontWeight.W500,
+      fontSize = 14.sp
+  ),
+  caption = TextStyle(
+      fontFamily = FontFamily.Default,
+      fontWeight = FontWeight.Normal,
+      fontSize = 12.sp
+  )
+  */
+)
