+++ shared/build.gradle.kts
@@ -47,10 +47,16 @@ kotlin {
 }
 
 android {
-    namespace = "com.example.mastadonclone"
+    namespace = "com.matrix159.mastadonclone.shared"
     compileSdk = 33
     defaultConfig {
         minSdk = 24
         targetSdk = 33
+    }
+
+    buildTypes {
+        getByName("release") {
+            isMinifyEnabled = true
+        }
     }
 }
