--- shared/src/androidMain/kotlin/com/example/mastadonclone/Platform.kt
@@ -1,7 +1,0 @@
-package com.example.mastadonclone
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${android.os.Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
