+++ shared/src/androidMain/kotlin/com/matrix159/mastadonclone/shared/Platform.kt
@@ -1,0 +1,8 @@
+package com.matrix159.mastadonclone.shared
+
+
+class AndroidPlatform : Platform {
+    override val name: String = "Android ${android.os.Build.VERSION.SDK_INT}"
+}
+
+actual fun getPlatform(): Platform = AndroidPlatform()
