--- shared/src/androidTest/kotlin/com/example/mastadonclone/androidTest.kt
@@ -1,12 +1,0 @@
-package com.example.mastadonclone
-
-import org.junit.Assert.assertTrue
-import org.junit.Test
-
-class AndroidGreetingTest {
-
-    @Test
-    fun testExample() {
-        assertTrue("Check Android is mentioned", Greeting().greet().contains("Android"))
-    }
-}
