+++ shared/src/androidTest/kotlin/com/matrix159/mastadonclone/androidTest.kt
@@ -1,0 +1,12 @@
+package com.matrix159.mastadonclone
+
+import org.junit.Assert.assertTrue
+import org.junit.Test
+
+class AndroidGreetingTest {
+
+    @Test
+    fun testExample() {
+        assertTrue("Check Android is mentioned", Greeting().greet().contains("Android"))
+    }
+}
