--- shared/src/commonMain/kotlin/com/example/mastadonclone/Greeting.kt
@@ -1,9 +1,0 @@
-package com.example.mastadonclone
-
-class Greeting {
-    private val platform: Platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
