--- shared/src/commonMain/kotlin/com/example/mastadonclone/Platform.kt
@@ -1,7 +1,0 @@
-package com.example.mastadonclone
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
