+++ shared/src/commonMain/kotlin/com/matrix159/mastadonclone/shared/Greeting.kt
@@ -1,0 +1,9 @@
+package com.matrix159.mastadonclone.shared
+
+class Greeting {
+    private val platform: Platform = getPlatform()
+
+    fun greet(): String {
+        return "Hello, ${platform.name}!"
+    }
+}
