+++ shared/src/commonMain/kotlin/com/matrix159/mastadonclone/shared/Platform.kt
@@ -1,0 +1,7 @@
+package com.matrix159.mastadonclone.shared
+
+interface Platform {
+    val name: String
+}
+
+expect fun getPlatform(): Platform
