--- shared/src/commonTest/kotlin/com/example/mastadonclone/commonTest.kt
@@ -1,12 +1,0 @@
-package com.example.mastadonclone
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class CommonGreetingTest {
-
-    @Test
-    fun testExample() {
-        assertTrue(Greeting().greet().contains("Hello"), "Check 'Hello' is mentioned")
-    }
-}
