+++ shared/src/commonTest/kotlin/com/matrix159/mastadonclone/shared/commonTest.kt
@@ -1,0 +1,12 @@
+package com.matrix159.mastadonclone.shared
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class CommonGreetingTest {
+
+    @Test
+    fun testExample() {
+        assertTrue(Greeting().greet().contains("Hello"), "Check 'Hello' is mentioned")
+    }
+}
