--- shared/src/iosMain/kotlin/com/example/mastadonclone/Platform.kt
@@ -1,9 +1,0 @@
-package com.example.mastadonclone
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
