+++ shared/src/iosMain/kotlin/com/matrix159/mastadonclone/shared/Platform.kt
@@ -1,0 +1,9 @@
+package com.matrix159.mastadonclone.shared
+
+import platform.UIKit.UIDevice
+
+class IOSPlatform: Platform {
+    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
+}
+
+actual fun getPlatform(): Platform = IOSPlatform()
