--- shared/src/iosTest/kotlin/com/example/mastadonclone/iosTest.kt
@@ -1,12 +1,0 @@
-package com.example.mastadonclone
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class IosGreetingTest {
-
-    @Test
-    fun testExample() {
-        assertTrue(Greeting().greet().contains("iOS"), "Check iOS is mentioned")
-    }
-}
