+++ shared/src/iosTest/kotlin/com/matrix159/mastadonclone/shared/iosTest.kt
@@ -1,0 +1,12 @@
+package com.matrix159.mastadonclone.shared
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class IosGreetingTest {
+
+    @Test
+    fun testExample() {
+        assertTrue(Greeting().greet().contains("iOS"), "Check iOS is mentioned")
+    }
+}
